build(codeql-analysis.yml): add CodeQL analysis workflow to the project
This commit adds a new workflow file to the project that sets up a CodeQL analysis job. The job is triggered on push and pull request events on the master branch, as well as on a weekly schedule. The job runs on an Ubuntu environment and uses a matrix strategy to analyze code written in C++. The workflow includes steps to checkout the repository, restore and install Vcpkg packages, configure CMake, and perform the CodeQL analysis.